Version 1.00
Here is version 1.0 of the Cooper RFCW5 scene-controller plug-in.

Please unzip this file and upload all 16 files into your Vera using apps->develop apps->luup files. Then create a new device in apps->develop apps->create device. In the "Upnp Device Filename" field, enter:
D_GenGenSceneControllerInstaller.xml
You can leave all other fields blank and then click on the "Create device" button.

This is a major update from the previous version which is actually a generic scene controller driver which supports both the Evolve LCD1 and the Cooper RFWC5. There is also some preliminary support for the Nexia One-Touch but this is not yet complete.

This has been tested with both the latest update of UI7 as well as the last UI5 version.

The new version fully supports both momentary, toggle, and 3 to 6-state buttons on the Cooper RFWC5 using both direct association as well as Vera scenes. The controller LEDs work as expected for toggle but will only light briefly for the 3 to 6-state modes.