Version 1.13
This version should be able to track Vera scenes in the indicator. (LEDs on Cooper RFWC5, higlighted lines on the Evolve LCD1)
You need to display the Screen page once in the Browser GUI for scene information to be available to the plug-in Lua code. You need to do this again any time that you change the scene.
Exclusive buttons should now work better. (only one exclusive can be active at a time.)
The class of devices that can be controlled and monitored is more general. Any device the support LoadLevelTarget or SwitchPower status should work.