Version 1.15
This version fixes several subtle bugs especially regarding configuring the Cooper RFWC5 and holding down the buttons for raise/lower. However it cannot resolve some of your basic problems whereby tapping one button and then holding down a different button causes strange behavior if either button is associated with a device which is not scene-capable.

Basically, the Cooper RFWC5 will (usually) only send Switch Multilevel Start/stop commands to scene-capable devices. To work around that, since version 1.14, the plug-in will forward such commands to the last device from which it received a scene activation or basic set command from. Unfortunately, if you hold down a different button, the RFWC5 will only send a pair of Multilevel Start/Stop commands to the controller and these command do not specify from which button they are coming from. In the case of tap off which sends a generic Basic Set off, this is handled by comparing the previous state of the indicator lights with the current state to decide which one you turned off. However, in the case of hold on/off, this trick is impossible. My plug-in assumes that you are holding down the same button that you last tapped. However, sometimes the controller will send the Multi-level start/stop messages to both Vera and the target device. (It is inconsistent, at least in my version of the RFWC5 firmware.) so both devices may actuate.

Furthermore, if a device is completely off (LED is off), holding down the button for a long time may cause the LED to flash, as if it was in some internal programming mode. Once my RFWC5 gets into this mode, it will stop responding to Indicator Get/Set commands until I power-cycle the RFWC5. This also looks to be a bug in the RFWC5 firmware. Your version may be newer than mine.

"Got CAN" errors are temporary hiccups in the Z-Wave protocol which can occur if Vera is sending a command just when a device is sending an unsolicited status message. In response, Vera waits for 1 second to let the other messages go through. Polling definitely exacerbates this problem. You can turn polling off in Vera global Z-Wave settings screen, but of course, you will not get updates from devices that don't support instant reporting. (My understanding is that Lutron holds a patent on instant reporting which was licensed by Cooper/Eaton, but not by other manufacturers.) The plug-in does not rely on polling per-se other than issue a poll command whenever you release a long-press on the RFWC5 to kick Vera to find out the final state of the target device after you let go of the button.

I highly recommend that you talk to your friend at Motion Control and have them update their Z-Wave firmware to support the SCENE_ACTIVATION and SCENE_ACTUATOR_CONFIGURATION command classes in their window covering motor.

The really bad problem is the "Failed to get lock" error and I'm still waiting for Vera Controls to fix this problem. Please complain to them as well. 

If you want to have the LED's track a Vera scene, you need to modify the configuration in the plug-in Screens/Configuration tab. When you change any settings related to a button in this tab. the UI will send a list of devices affected by the scene to the Lua code. (Unfortunately, Lua cannot see the details of Vera scenes.)

We are quickly getting to the point of diminishing returns where we are discovering quirks and limitations in these devices. You are pushing the RFWC5 hard and trying to make it do things that it was never designed to do. It is natural that we will start finding bugs and other quirks in firmware that we don't have control over.